// src/ui/app/root/application/models/Assessment/LevelProgramClassMap.ts
export interface ILevelProgramClassMap {
    levelProgramClassId: string;
    levelId: string;
    programId: string;
    classId: string;
    statusId: number;
}

export interface IVWLevelProgramClassMap {
    levelProgramClassId: string;
    levelName: string;
    programName: string;
    className: string;
    statusId: number;
}